Completed implementation of this class
Added actions to perform for each key event for the Reply packet type. For the Join packet type, added in actions that allow the RemoteClient to join the Mazewar session. For the Bye packet type, added in actions that remove the client from the hash map and maze, and if the client is a GUIClient, exits the Mazewar game.